refactor: use context to store user information
